AugSynth: include predictors in weights optimization
Augsynth was not using other predictors besides the outcome variable in the weights optimization. The ridge regression is not enough to adjust the weights based on other (or special) predictors. Including the predictors in the initial weights optimization ensures the treatment and synthetic variable is almost equal in all predictors in the pre-treatment period, where possible.