Add stopping condition for while loop
The while loop in robust.py assumes that we've specified a high enough threshold to end. However, if the threshold is low enough, we will continue looping until we try to access an element that is beyond the size of s, creating an error. We have now added a stopping condition to that.